requirements.txt
Flask==3.0.3
Flask-MySQLdb==2.0.0
mysqlclient==2.2.4
python-dotenv==1.0.1
Werkzeug==3.0.3
